requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
astroid==2.15.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
botocore==1.29.111
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contextlib2==21.6.0
dill==0.3.6
et-xmlfile==1.1.0
fastapi==0.95.0
fastapi-cache2==0.2.1
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.1
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
kink==0.6.6
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pendulum==2.1.2
platformdirs==3.2.0
pluggy==1.0.0
pony==0.7.16
protobuf==4.22.1
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PyJWT==2.6.0
pylint==2.17.2
pytest==7.3.1
pytest-schema==0.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0
redis==4.5.5
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
schema==0.7.5
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
yarl==1.9.2
